% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_distinct_colours.R
\name{generate_distinct_colours}
\alias{generate_distinct_colours}
\title{Generate Distinct Colours via the hsv function}
\usage{
generate_distinct_colours(n)
}
\arguments{
\item{n}{The number of hex codes to generate}
}
\value{
A vector of hex codes of length, \strong{n}
}
\description{
Creates a vector of (not necessarily) unique colours - that are not to dark and not too light.
}
\examples{

## Not Run:
generate_distinct_colours(3)
# "#757B48" "#2117C3" "#871C0D"

}
\seealso{
\code{\link{hsv}}
}
