% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_functions}
\alias{load_functions}
\title{Load all R functions}
\usage{
load_functions(path = ".", pattern = NULL, all.files = FALSE,
  full.names = FALSE, recursive = FALSE, ignore.case = FALSE,
  include.dirs = FALSE, no.. = FALSE, ...)
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to the working directory, getwd(). Tilde expansion (see path.expand) is performed. Missing values will be ignored.}

\item{pattern}{an optional regular expression. Only file names which match the regular expression will be returned.}

\item{all.files}{a logical value. If FALSE, only the names of visible files are returned. If TRUE, all file names will be returned.}

\item{full.names}{a logical value. If TRUE, the directory path is prepended to the file names to give a relative file path. If FALSE, the file names (rather than paths) are returned.}

\item{recursive}{logical. Should the listing recurse into directories?}

\item{ignore.case}{logical. Should pattern-matching be case-insensitive?}

\item{include.dirs}{logical. Should subdirectory names be included in recursive listings? (They always are in non-recursive ones).}

\item{no..}{logical. Should both "." and ".." be excluded also from non-recursive listings?}

\item{...}{Other options pased to \code{\link{source}}}
}
\description{
Load R functions from files in a directory to the global environment
}
