% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\alias{\%inx\%}
\alias{\%notinx\%}
\title{Value matching}
\usage{
x \%notin\% y

x \%inx\% y

x \%notinx\% y
}
\arguments{
\item{x}{vector or NULL: the values to be matched. Long vectors are supported.}

\item{y}{vector or NULL: the values to be matched against. Long vectors are not supported.}
}
\value{
\code{\%notin\%} A logical vector, indicating if a match was not located for each element of \code{x}: thus the values are TRUE or FALSE and never NA.

\code{\%inx\%} returns a sub-vector of \code{x} with the values that were found in \code{y}.

\code{\%notinx\%} returns a sub-vector of \code{x} with the values that were not found in \code{y}.
}
\description{
\code{\%in\%} returns logical vector indicating values that do not have a match.
\code{\%inx\%} returns a vector of the values that have a match.
\code{\%notinx\%} returns a vector of the values that do not have a match.
}
\examples{
vals <- c("a", "notin", 3)
vals \%notin\% letters
vals \%inx\% letters
vals \%notinx\% letters
}
\seealso{
\code{\link[base:match]{base::match()}}
}
